changed duration

+++ search-microservice/src/main/java/com/stackroute/matchmaker/searchmicroservice/model/Result.java
@@ -15,7 +15,7 @@ import lombok.Data;
 public class Result {
 	private String profileId;
 	private List<String> skills;
-	private long experience;
+	private int experience;
 	private String city;
 
 }
